HW2

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,14 +1,18 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
-import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
@@ -25,22 +29,41 @@ public class InMemoryMealRepository implements MealRepository {
             return meal;
         }
         // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        if (checkUser(meal.getId())) {
+            return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        }
+        return null;
     }
 
     @Override
     public boolean delete(int id) {
-        return repository.remove(id) != null;
+        if (checkUser(id)) {
+            return repository.remove(id) != null;
+        }
+        return false;
     }
+
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        if (checkUser(id))
+            return repository.get(id);
+        else return null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public List<Meal> getAll() {
+        if (repository != null) {
+            throw new NullPointerException("repository is empty");
+        }
+        List<Meal> result = (List<Meal>) repository.values();
+        Collections.sort(result);
+        return result;
+    }
+
+    @Override
+    public boolean checkUser(int id) {
+        return SecurityUtil.authUserId() == repository.get(id).getUserId() && id != 0;
     }
 }
